load orderList

+++ app/src/main/java/com/example/yogiyo_clone/src/main/orderlist/models/response/Store.kt
@@ -13,7 +13,7 @@ data class Store(
     val isExpress: String?,
     val isSuperRedweek: String?,
     val limitCharge: Int,
-    val rateAvg: Int?,
+    val rateAvg: Float?,
     val storeIdx: Int,
     val title: String
 )
